+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 Yii Framework 2 smarty extension Change Log
 ===========================================
 
-2.0.9 under development
+2.0.9 November 19, 2019
 -----------------------
 
 - Enh #36: Added `{js}` function allowing instantiate `yii\web\JsExpression` (leandrogehlen)
+
 
 2.0.8 April 16, 2019
 --------------------